Int types instead of size_t and uint8_t in spectre.c

+++ benchmarks/spectre/spectre.c
@@ -11,17 +11,17 @@ extern int __VERIFIER_nondet_int(void);
 #define SIZE    (1)
 
 unsigned int array1_size = 16;
-uint8_t array1[16] = { 1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16 };
-uint8_t array2[256 * SIZE];
+int array1[16] = { 1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16 };
+int array2[256 * SIZE];
 
 char * secret = "The Magic Words are Squeamish Ossifrage.";
 
-uint8_t temp = 0; /* Used so compiler won't optimize out victim_function() */
+int temp = 0; /* Used so compiler won't optimize out victim_function() */
 
 // ----------------------------------------------------------------------------------------
 // EXAMPLE 1:  This is the sample function from the Spectre paper.
 // ----------------------------------------------------------------------------------------
-void victim_function_v01(size_t x) {
+void victim_function_v01(int x) {
     if (x < array1_size) {
         temp &= array2[array1[x] * SIZE];
     }
@@ -30,8 +30,8 @@ void victim_function_v01(size_t x) {
 // ----------------------------------------------------------------------------------------
 // EXAMPLE 2:  Moving the leak to a local function that can be inlined.
 // ----------------------------------------------------------------------------------------
-void leakByteLocalFunction(uint8_t k) { temp &= array2[(k)* SIZE]; }
-void victim_function_v02(size_t x) {
+void leakByteLocalFunction(int k) { temp &= array2[(k)* SIZE]; }
+void victim_function_v02(int x) {
     if (x < array1_size) {
         leakByteLocalFunction(array1[x]);
     }
@@ -43,8 +43,8 @@ void victim_function_v02(size_t x) {
 // Comments: Output is unsafe.  The same results occur if leakByteNoinlineFunction()
 // is in another source module.
 // ----------------------------------------------------------------------------------------
-__declspec(noinline) void leakByteNoinlineFunction(uint8_t k) { temp &= array2[(k)* SIZE]; }
-void victim_function_v03(size_t x) {
+__declspec(noinline) void leakByteNoinlineFunction(int k) { temp &= array2[(k)* SIZE]; }
+void victim_function_v03(int x) {
     if (x < array1_size) {
         leakByteNoinlineFunction(array1[x]);
     }
@@ -55,7 +55,7 @@ void victim_function_v03(size_t x) {
 //
 // Comments: Output is unsafe.
 // ----------------------------------------------------------------------------------------
-void victim_function_v04(size_t x) {
+void victim_function_v04(int x) {
     if (x < array1_size) {
         temp &= array2[array1[x << 1] * SIZE];
     }
@@ -66,7 +66,7 @@ void victim_function_v04(size_t x) {
 //
 // Comments: Output is unsafe.
 // ----------------------------------------------------------------------------------------
-void victim_function_v05(size_t x) {
+void victim_function_v05(int x) {
      int i;
      if (x < array1_size) {
           for (i = x - 1; i >= 0; i--)
@@ -79,8 +79,8 @@ void victim_function_v05(size_t x) {
 //
 // Comments: Output is unsafe.
 // ----------------------------------------------------------------------------------------
-uint8_t array_size_mask = 15;
-void victim_function_v06(size_t x) {
+int array_size_mask = 15;
+void victim_function_v06(int x) {
     if ((x & array_size_mask) == x) {
         temp &= array2[array1[x] * SIZE];
     }
@@ -91,8 +91,8 @@ void victim_function_v06(size_t x) {
 //
 // Comments: Output is unsafe.
 // ----------------------------------------------------------------------------------------
-void victim_function_v07(size_t x) {
-    static size_t last_x = 0;
+void victim_function_v07(int x) {
+    static int last_x = 0;
     if (x == last_x) {
         temp &= array2[array1[x] * SIZE];
     }
@@ -104,7 +104,7 @@ void victim_function_v07(size_t x) {
 // ----------------------------------------------------------------------------------------
 // EXAMPLE 8:  Use a ?: operator to check bounds.
 // ----------------------------------------------------------------------------------------
-void victim_function_v08(size_t x) {
+void victim_function_v08(int x) {
     temp = array2[array1[x < array1_size ? (x + 1) : 0] * SIZE];
 }
 
@@ -113,7 +113,7 @@ void victim_function_v08(size_t x) {
 //
 // Comments: Output is unsafe.
 // ----------------------------------------------------------------------------------------
-void victim_function_v09(size_t x, int *x_is_safe) {
+void victim_function_v09(int x, int *x_is_safe) {
     if (*x_is_safe) {
         temp &= array2[array1[x] * SIZE];
     }
@@ -127,7 +127,7 @@ void victim_function_v09(size_t x, int *x_is_safe) {
 // array1[x] == k.  If so, the victim reads from array2[0].  The attacker can try
 // values for k until finding the one that causes array2[0] to get brought into the cache.
 // ----------------------------------------------------------------------------------------
-void victim_function_v10(size_t x, uint8_t k) {
+void victim_function_v10(int x, int k) {
     if (x < array1_size) {
         if (array1[x] == k) {
             temp &= array2[0];
@@ -140,7 +140,7 @@ void victim_function_v10(size_t x, uint8_t k) {
 //
 // Comments: Output is unsafe.
 // ----------------------------------------------------------------------------------------
-int mymemcmp(const void *cs, const void *ct, size_t count)
+int mymemcmp(const void *cs, const void *ct, int count)
 {
     const unsigned char *su1, *su2;
     int res = 0;
@@ -152,7 +152,7 @@ int mymemcmp(const void *cs, const void *ct, size_t count)
     }
     return res;
 }
-void victim_function_v11(size_t x) {
+void victim_function_v11(int x) {
     if (x < array1_size) {
         temp &= mymemcmp(&temp, array2 + (array1[x] * SIZE), 1);
     }
@@ -163,7 +163,7 @@ void victim_function_v11(size_t x) {
 //
 // Comments: Output is unsafe.
 // ----------------------------------------------------------------------------------------
-void victim_function_v12(size_t x, size_t y) {
+void victim_function_v12(int x, int y) {
     if ((x + y) < array1_size) {
         temp &= array2[array1[x + y] * SIZE];
     }
@@ -174,13 +174,13 @@ void victim_function_v12(size_t x, size_t y) {
 //
 // Comments: Output is unsafe. We removed the "__inline" since anyways BMC inlines all the calls
 // ----------------------------------------------------------------------------------------
-int is_x_safe(size_t x) {
+int is_x_safe(int x) {
     if (x < array1_size) {
         return 1;
     }
     return 0;
 }
-void victim_function_v13(size_t x) {
+void victim_function_v13(int x) {
     if (is_x_safe(x)) {
         temp &= array2[array1[x] * SIZE];
     }
@@ -191,7 +191,7 @@ void victim_function_v13(size_t x) {
 //
 // Comments: Output is unsafe.
 // ----------------------------------------------------------------------------------------
-void victim_function_v14(size_t x) {
+void victim_function_v14(int x) {
     if (x < array1_size) {
         temp &= array2[array1[x ^ 255] * SIZE];
     }
@@ -202,7 +202,7 @@ void victim_function_v14(size_t x) {
 //
 // Comments: Output is unsafe.
 // ----------------------------------------------------------------------------------------
-void victim_function_v15(size_t *x) {
+void victim_function_v15(int *x) {
     if (*x < array1_size) {
         temp &= array2[array1[*x] * SIZE];
     }
@@ -210,8 +210,8 @@ void victim_function_v15(size_t *x) {
 
 int main()
 {
-    size_t x = __VERIFIER_nondet_int();
-    size_t k = __VERIFIER_nondet_int();
+    int x = __VERIFIER_nondet_int();
+    int k = __VERIFIER_nondet_int();
     #ifdef v01
     victim_function_v01(x);
     #endif
